+++ src/main/java/ca/on/oicr/pineryreports/util/GeneralUtils.java
@@ -1,16 +1,33 @@
 package ca.on.oicr.pineryreports.util;
 
-import java.time.format.DateTimeFormatter;
-import java.time.temporal.TemporalAccessor;
 import java.util.function.Predicate;
+import java.util.regex.Matcher;
+import java.util.regex.Pattern;
 
 import ca.on.oicr.ws.dto.RunDto;
 
 public class GeneralUtils {
 
-  public static String timeStringToYyyyMmDd(String dateTime) {
-    TemporalAccessor temp = DateTimeFormatter.ofPattern("yyyy-MM-dd'T'HH:mm:ss'Z'").parse(dateTime);
-    return DateTimeFormatter.ofPattern("yyyy-MM-dd").format(temp);
+  private GeneralUtils() {
+    throw new IllegalStateException("Util class not intended for instantiation");
+  }
+
+  /**
+   * Removes the time portion of a date/time String
+   * 
+   * @param datetime
+   *          String in format "YYYY-MM-DD..."
+   * @return String in format "YYYY-MM-DD" (anything beyond this is truncated)
+   */
+  public static String removeTime(String datetime) {
+    if (datetime == null) {
+      return null;
+    }
+    Matcher m = Pattern.compile("^(\\d{4}-\\d{2}-\\d{2}).*").matcher(datetime);
+    if (!m.matches()) {
+      throw new IllegalArgumentException("Datetime string is not in expected format (YYYY-MM-DD...)");
+    }
+    return m.group(1);
   }
   
   public static Predicate<RunDto> byEndedBetween(String start, String end) {
