+++ src/main/java/ca/on/oicr/pineryreports/util/SampleUtils.java
@@ -7,15 +7,13 @@ import java.util.List;
 import java.util.Map;
 import java.util.Set;
 import java.util.function.Predicate;
-import java.util.regex.Matcher;
-import java.util.regex.Pattern;
 import java.util.stream.Collectors;
+
+import com.google.common.collect.Lists;
 
 import ca.on.oicr.ws.dto.AttributeDto;
 import ca.on.oicr.ws.dto.SampleDto;
 import ca.on.oicr.ws.dto.SampleReferenceDto;
-
-import com.google.common.collect.Lists;
 
 public class SampleUtils {
 
@@ -206,23 +204,6 @@ public class SampleUtils {
     df.setMinimumFractionDigits(decimalPlaces);
     df.setGroupingUsed(false);
     return df.format(number);
-  }
-  
-  /**
-   * Removes the time portion of a date/time String
-   * 
-   * @param datetime String in format "YYYY-MM-DD..."
-   * @return String in format "YYYY-MM-DD" (anything beyond this is truncated)
-   */
-  public static String removeTime(String datetime) {
-    if (datetime == null) {
-      return null;
-    }
-    Matcher m = Pattern.compile("^(\\d{4}-\\d{2}-\\d{2}).*").matcher(datetime);
-    if (!m.matches()) {
-      throw new IllegalArgumentException("Datetime string is not in expected format (YYYY-MM-DD...)");
-    }
-    return m.group(1);
   }
   
   public static boolean isRnaLibrary(SampleDto library, Map<String, SampleDto> mapById) {
